01-python-crud-with-k8s-deploy/requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
google-cloud-pubsub==2.9.0